Implements basic link communication
Hosts may now send packets over links.  Link transport should be working
in the general case; dynamic cost may not be correctly implemented,
however.  Also renamed resource module because of a naming conflict with
a module in the Python standard library.  Some functionality was moved
from the resources module, as it became more appropriate to include it
in process.