Sparse unacknowledged packet container for flows
Instead of storing None values at the id of an acknowedged packet in a
list containing one entry per packet, Flow now has a dictionary of
unacknowledged packets (mapping packet ID -> packet instance), which is
a sparse version of the previous mapping, in which list index was equal
to packet id.  Also added some docstrings & comments to Flow.  I want to
have program execution logged (to stderr) so that we can watch a live
simulation easily, but for now, I just added a logger for the process
module & wrote a TODO.