Minor bugfix, barrier updates
Changed the network class to actually use MonitoredEnvironment, and
fixed some scope errors in MonitoredEnvironment.  I also made a minor
change to the way that updates are performed; now, values are only
updated at the end of a time period.  Thus, if multiple "simultaneous"
processes modify the same value, it is only updated once, after all
events in the present are finished (potentially modifying values)