Fixes documentation build
The documentation for some classes was not building correctly, since
they derived from modules which were mocked in the Sphinx configuration
(for Read The Docs).  I put the RtD build into a virtualenv, and added a
requirements.txt with the previously mocked modules, so that the
documentation will build correctly once these changes get pushed to
develop.