Re-implements TCP Reno, minor bugfix/style changes
After merging the TCP-Reno branch, I found that TCP Reno was not working
(it seemed to have been built on an old architecture, and not updated
correctly), and that the existing code was too far out of date to use.
I ended up re-implementing TCP Reno, as it took less time than trying to
fix the old code.  Made some stylistic changes to the Router class, and
in doing so found that Router.disconnect was out of date, so I updated
it.  Also fixed a minor error in HostResource._receive