Updates Network constructor to support TCP Reno
I added a keyword argument to the Network constructor to allow users to
specify which TCP algorithm to use, and updated the docstring to reflect
the changes.  One limitation to note is that all flows in the test case
will use the specified TCP algorithm; there is currently no way
(besides specifying them manually in the gui/with an adjacency list) to
mix different TCP algorithms.