Fixes TCP Reno fast recovery
TCP Reno was getting stuck in fast recovery mode, and it turned out to
be an issue with the timeout length (it was timing out during fast
recovery, causing it to re-enter fast recovery mode, creating an
inescapable loop).  I fixed it by setting the timeout to be the maximum
of the current timeout length and 2 * RTT each time Reno receives an
acknowledgement.