Finishes clean up of process module
I added the Router.routers property to simplify accessing the list of
connected routers, some inline comments and the _flush_wait attribute (to
get rid of the magic number in Link._flush for the empty buffer timeout)
in the Link class.  Also updated Link.cost to use the now directional
LinkBuffer.buffered, and moved the call to LinkBuffered.update_buffered
to Link.receive, i.e., when the buffer increases in size.