Fixes over capacity link transmission
Because we were calculating transmission as the time-averaged sum of all
traffic on the links, we were getting transmission rates greater than
the link capacity.  Transmission is now calculated directionally, and
the graphs look correct.  Also made some changes to the _reset method to
support this feature, as well as some minor stylistic
changes/documentation updates