Adds tag-based filtering to the simulate CLI
I added options to the simulate script to allow users to specify which
links, hosts, and flows they wish to display data for.  If none are
specfied, all data will be displayed.  There is still no option to
control which monitored values are output, but that should probably be
controlled from within the simulation code (we have to run the entire
simulation anyways, why display less data than we generated)